# openapi.yaml
openapi: 3.1.0
info:
  title: Autonomous AI Agent Safety & Permission Control System
  description: Local middleware firewall intercepting agent actions, scoring risk,
    and enforcing policy.
  version: 1.0.0
paths:
...
  /api/v1/models/emergency-retire:
    post:
      tags:
      - Model Governance
      summary: Emergency Retire
      description: Immediately retire the active model and fallback to known-good
        or lockdown.
      operationId: emergency_retire_api_v1_models_emergency_retire_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Emergency Retire Api V1 Models Emergency Retire Post
      security:
      - OAuth2PasswordBearer: []
  /api/v1/security/lockdown:
    post:
      tags:
      - Security Operations
      summary: Trigger Lockdown
      description: Trigger a global system lockdown. Fails closed all evaluate-action
        operations.
      operationId: trigger_lockdown_api_v1_security_lockdown_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Trigger Lockdown Api V1 Security Lockdown Post
      security:
      - OAuth2PasswordBearer: []
  /api/v1/security/unlock:
    post:
      tags:
      - Security Operations
      summary: Release Lockdown
      description: Release system lockdown and restore normal safety checks.
      operationId: release_lockdown_api_v1_security_unlock_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Release Lockdown Api V1 Security Unlock Post
      security:
      - OAuth2PasswordBearer: []
  /api/v1/security/status:
    get:
      tags:
      - Security Operations
      summary: Get Lockdown Status
      description: Retrieve the current lockdown state of the security platform.
      operationId: get_lockdown_status_api_v1_security_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Lockdown Status Api V1 Security Status Get
      security:
      - OAuth2PasswordBearer: []
  /api/v1/security/audit/verify:
    get:
      tags:
      - Security Operations
      summary: Verify Audit Logs
      description: Verify cryptographic integrity of the security audit log chain.
      operationId: verify_audit_logs_api_v1_security_audit_verify_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Verify Audit Logs Api V1 Security Audit Verify Get
      security:
      - OAuth2PasswordBearer: []
  /api/v1/security/audit/export:
    get:
      tags:
      - Security Operations
      summary: Export Audit Logs
      description: Export paginated compliance-audited security events.
      operationId: export_audit_logs_api_v1_security_audit_export_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: start_time
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ISO format start datetime filter
          title: Start Time
        description: ISO format start datetime filter
      - name: end_time
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ISO format end datetime filter
          title: End Time
        description: ISO format end datetime filter
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Severity category filter
          title: Severity
        description: Severity category filter
      - name: event_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Event classification filter
          title: Event Type
        description: Event classification filter
      - name: agent_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Target agent filter
          title: Agent Id
        description: Target agent filter
      - name: incident_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Linked incident filter
          title: Incident Id
        description: Linked incident filter
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page index
          default: 1
          title: Page
        description: Page index
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Items per page
          default: 100
          title: Limit
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Export Audit Logs Api V1 Security Audit Export Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/incidents:
    get:
      tags:
      - Incident Response
      summary: List Incidents
      description: Retrieve all logged security incidents.
      operationId: list_incidents_api_v1_incidents_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IncidentResponse'
                type: array
                title: Response List Incidents Api V1 Incidents Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - Incident Response
      summary: Create Incident
      description: Declare a new security operational incident.
      operationId: create_incident_api_v1_incidents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/incidents/{incident_id}:
    get:
      tags:
      - Incident Response
      summary: Get Incident
      description: Get detail specifications of a single security incident.
      operationId: get_incident_api_v1_incidents__incident_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Incident Response
      summary: Update Incident
      description: Modify details, ownership, or status of an incident.
      operationId: update_incident_api_v1_incidents__incident_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/incidents/{incident_id}/resolve:
    post:
      tags:
      - Incident Response
      summary: Resolve Incident
      description: Transition an incident state to resolved.
      operationId: resolve_incident_api_v1_incidents__incident_id__resolve_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Notes
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/agents/register:
    post:
      summary: Register Agent
      description: 'Register a new AI agent and issue a plaintext API key.


        The plaintext key is returned ONCE.  The caller must store it securely and

        pass it as the X-Agent-Key header on every /evaluate-action call.


        Idempotent: if the agent_id already exists and is already registered (has
        a

        key hash), returns 409 Conflict.


        The assigned role must be one of: ResearchAgent, DeveloperAgent, OperationsAgent.

        The role cannot be changed after registration (Phase 3 will add management
        APIs).'
      operationId: register_agent_api_v1_agents_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegisterResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/evaluate-action:
    post:
      summary: Evaluate Action
      description: "Core firewall evaluation endpoint.\n\nPipeline order (Phase 2):\n\
        \    1. Authentication (X-Agent-Key validation) \u2190 NEW\n    2. Rate limiting\
        \ (Redis counter)\n    3. Agent DB get-or-create\n    4. ML Risk Engine (SentenceTransformer\
        \ + XGBoost)\n    5. Policy Engine (RBAC)\n    6. PostgreSQL persistence\n\
        \    7. WebSocket broadcast (if human approval required)\n\nAUTH_ENABLED=true:\n\
        \    - X-Agent-Key is mandatory.\n    - agent_id and role are derived from\
        \ the AUTHENTICATED credential.\n    - Self-reported agent_id in the request\
        \ body is IGNORED for identity.\n      (It is still used to produce useful\
        \ audit log entries; the actual\n       authoritative identity is the credential-bound\
        \ agent.)\n\nAUTH_ENABLED=false (Phase 1 compatibility):\n    - X-Agent-Key\
        \ is optional.\n    - agent_id and role fall back to self-reported + substring\
        \ matching."
      operationId: evaluate_action_api_v1_evaluate_action_post
      parameters:
      - name: X-Agent-Key
        in: header
        required: false
        schema:
          type: string
          title: X-Agent-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionEvaluationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionEvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ActionEvaluationRequest:
      properties:
        agent_id:
          type: string
          title: Agent Id
          description: Unique identifier for the AI agent
        action:
          type: string
          title: Action
          description: Action name or type being attempted
        parameters:
          type: object
          title: Parameters
          description: Parameters associated with the action
        requested_at:
          type: string
          title: Requested At
          description: ISO-8601 timestamp of the request
      type: object
      required:
      - agent_id
      - action
      - requested_at
      title: ActionEvaluationRequest
    ActionEvaluationResponse:
      properties:
        decision:
          type: string
          title: Decision
          description: 'Decision: allow | block | require_human_approval'
        risk_level:
          type: string
          title: Risk Level
          description: 'Risk Level: low | medium | high'
        reason:
          type: string
          title: Reason
          description: Explanation of the evaluation decision
        model_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Version
        feature_schema_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Feature Schema Version
        policy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Policy Version
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      type: object
      required:
      - decision
      - risk_level
      - reason
      title: ActionEvaluationResponse
...
    AgentRegisterRequest:
      properties:
        agent_id:
          type: string
          title: Agent Id
          description: Unique identifier for the AI agent
        name:
          type: string
          title: Name
          description: Human-readable agent name
        role:
          type: string
          title: Role
          description: 'RBAC role: ResearchAgent | DeveloperAgent | OperationsAgent'
      type: object
      required:
      - agent_id
      - name
      - role
      title: AgentRegisterRequest
    AgentRegisterResponse:
      properties:
        agent_id:
          type: string
          title: Agent Id
        role:
          type: string
          title: Role
        api_key:
          type: string
          title: Api Key
          description: "Plaintext API key \u2014 shown ONCE. Store it securely. Pass\
            \ it in X-Agent-Key header."
        message:
          type: string
          title: Message
      type: object
      required:
      - agent_id
      - role
      - api_key
      - message
      title: AgentRegisterResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IncidentCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 3
          title: Title
        description:
          type: string
          minLength: 5
          title: Description
        severity:
          type: string
          pattern: ^(low|medium|high|critical)$
          title: Severity
          default: medium
        affected_agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Agent Id
        affected_model_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Model Version
        affected_policy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Policy Version
      type: object
      required:
      - title
      - description
      title: IncidentCreate
    IncidentResponse:
      properties:
        incident_id:
          type: string
          title: Incident Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        severity:
          type: string
          title: Severity
        status:
          type: string
          title: Status
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
        resolved_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolved At
        created_by:
          type: string
          title: Created By
        assigned_to:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned To
        affected_agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Agent Id
        affected_model_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Model Version
        affected_policy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Affected Policy Version
        resolution_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution Notes
      type: object
      required:
      - incident_id
      - title
      - description
      - severity
      - status
      - created_at
      - updated_at
      - created_by
      title: IncidentResponse
    IncidentUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        severity:
          anyOf:
          - type: string
            pattern: ^(low|medium|high|critical)$
          - type: 'null'
          title: Severity
        status:
          anyOf:
          - type: string
            pattern: ^(open|investigating|contained|resolved|closed)$
          - type: 'null'
          title: Status
        assigned_to:
          anyOf:
          - type: string
          - type: 'null'
          title: Assigned To
        resolution_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution Notes
      type: object
      title: IncidentUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/operator/login